+++ lib/flutter_widgetz.dart
@@ -30,6 +30,7 @@ part 'src/countdown.dart';
 part 'src/counter_field.dart';
 part 'src/date_field.dart';
 part 'src/date_range_field.dart';
+part 'src/directional_pad.dart';
 part 'src/divider.dart';
 part 'src/drawer.dart';
 part 'src/dropdown_field.dart';